api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic==0.40.0
PyGithub==2.5.0
pydantic==2.10.0
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-dotenv==1.0.0
supabase==2.10.0
jinja2==3.1.4
asyncpg==0.30.0
pyyaml==6.0.2
